Changed currency representation from float to int, added public and private keys

+++ coursework1/src/SharedObjects/Bid.java
@@ -3,9 +3,9 @@
 */
 public class Bid
 {
-    public float bidPrice;
+    public int bidPrice;
     public Account bidder;
-    public Bid(float price, Account bidderAccount)
+    public Bid(int price, Account bidderAccount)
     {
         bidder = bidderAccount;
         bidPrice = price;
